+++ lib/tappay/credit_card/instalment.rb
@@ -1,17 +1,53 @@
 module Tappay
   module CreditCard
-    class InstalmentBase < PayBase
+    class InstalmentBase < Client
       def initialize(options = {})
         super
         validate_instalment_options!
       end
 
+      def execute
+        post(endpoint_url, payment_data)
+      end
+
       protected
 
+      def endpoint_url
+        raise NotImplementedError, "Subclass must implement abstract method 'endpoint_url'"
+      end
+
       def payment_data
-        super.merge(
+        {
+          partner_key: Tappay.configuration.partner_key,
+          merchant_id: options[:merchant_id] || Tappay.configuration.merchant_id,
+          amount: options[:amount],
+          details: options[:details],
+          currency: options[:currency] || 'TWD',
+          order_number: options[:order_number],
+          redirect_url: options[:redirect_url],
+          three_domain_secure: options[:three_domain_secure] || false,
+          card_holder: card_holder_data,
           instalment: options[:instalment]
-        )
+        }
+      end
+
+      def card_holder_data
+        return unless options[:card_holder]
+
+        case options[:card_holder]
+        when CardHolder
+          options[:card_holder].to_h
+        when Hash
+          options[:card_holder]
+        else
+          raise ValidationError, "Invalid card_holder format"
+        end
+      end
+
+      def validate_options!
+        required = [:amount, :details]
+        missing = required.select { |key| options[key].nil? }
+        raise ValidationError, "Missing required options: #{missing.join(', ')}" if missing.any?
       end
 
       def validate_instalment_options!
@@ -21,7 +57,7 @@ module Tappay
       end
     end
 
-    class Instalment < PayBase
+    class Instalment
       def self.by_prime(options = {})
         InstalmentByPrime.new(options)
       end